Started TDDing out a new design for building sitemaps as well as a sitemap index file

+++ Source/BusinessLogic/Jobs/SitemapGenerator/RegenerateSitemapsJobResult.cs
@@ -2,7 +2,7 @@
 {
     public class RegenerateSitemapsJobResult
     {
-        public int TimeElapsedInMilliseconds { get; set; }
+        public long TimeElapsedInMilliseconds { get; set; }
         public int NumberOfSitemapsGenerated { get; set; }
     }
 }
